chromedb: add support for database versions > 23
Recent database versions have updated the format of encrypted cookies.
Prior to encryption, the payload is prefixed with a SHA256 digest of the domain
("host_key") of the cookie, and Chrome checks for these digests when reading
the cookies in.

Ensure we handle this prefix when reading and writing cookies in versions of
the database that require it.